# Refactor how `CowData` manages its allocations