--- a/core/typedefs.h
+++ b/core/typedefs.h
@@ -127,6 +127,11 @@ constexpr auto CLAMP(const T m_a, const T2 m_min, const T3 m_max) {
 	return m_a < m_min ? m_min : (m_a > m_max ? m_max : m_a);
 }
 
+template <typename T>
+constexpr auto ALIGN_AT(const T m_number, const T m_alignment) {
+	return ((m_number + (m_alignment - 1)) / m_alignment) * m_alignment;
+}
+
 // Generic swap template.
 #ifndef SWAP
 #define SWAP(m_x, m_y) __swap_tmpl((m_x), (m_y))