Context: I was looking into making CowData use 64-bit indices to solve the various crashes / issues caused by too large arrays, and found that the internal structure of CowData is fundamentally not suited for such a thing (and also really weird) so I went and reworked how the internal data storage is managed.

To the outside this PR shouldn't change any behavior, except that allocating somewhat larger arrays works now (and also on MSVC there are no longer cases where operations return OK but actually failed).

There are a few more improvements that could / should be done, but I stopped before feature creep completely overtook this.

Needs https://github.com/godotengine/godot/pull/74555 to pass CI.

- Use an explicit type for the data prefix to enable easier changes.
- Explicitly store the capacity instead of using the next power of two.
- If resizing to a significantly larger size (more 2x the current capacity), assume that the caller has just resized to their final size and don't allocate any excess capacity.
- Trim the capacity less aggressively to fix edge cases when repeatedly adding and removing elements causing extreme performance issues.
- Do not rely on overflow checked operations to determine the maximum size, to avoid crashes with some compilers when using too large sizes.
- On 64-bit system increase the the maximum size to 0x7ffffff0, generally by a factor of the size of the elements.
- On 32-bit system the maximum size remains limited by the size of the address space / memory, but in some cases it may still be possible to allocate larger arrays compared to before.
- Add a minimum capacity of 4 elements to remove some unnecessary reallocation for small arrays.
- Remove a few unnecessary potential copies when resizing, adding or removing elements.

---

What to do with the `ERR_PROPAGATE` macro?
I added this to check an `Error` return value and to return it from the current method if it is not OK, is this usefull for other places in the engine as well and should be added to the other ERR macros? Should I just inline that check everywhere? Or can it just stay as is?

---

What about alignment?
The old version of this just didn't care about alignment and just returned what it got from the allocator (which is malloc aligned + 16 bytes).
This assumes that the allocator pointer is aligned correctly and then offsets the first element by an multiple of the alignment, but I am not sure if the alignment is correct for types with large alignments, ie simd types. Should this be extended to ensure correct alignment for such types?

---

About the buffers in `max_size`:
The 64 bytes buffer to filling the entire address space with one array is arbitrary and not going to matter beyond preventing overflows when computing the size of the allocation as way more than that are going to be filled with godot, so such allocations are going to fail anyways.
The 15 indices buffer is similarly arbitrary and serves to prevent overflows when computing `size()+1` and similar things.
For reference in C# the maximum array size is `0x7fffffc7` bytes regardless of element size. (Buffer of 56).

---

About that performance hole in the current implementation:
When repeatedly increasing and decreasing the size of the allocation across a power of 2, on every operation the entire backing array will be reallocated, causing (in constructed examples 1000x+) slowdowns compared to doing the same with just a few more or less elements.
```gdscript
func _ready():
	perf(43689)
	perf(43690)
	perf(43691)
func perf(size):
	var start = Time.get_ticks_msec()
	var arr = []
	arr.resize(size)
	for i in range(10000):
		arr.push_back(null)
		arr.pop_back()
	var end = Time.get_ticks_msec()
	print("%d took %dms" % [size, end - start])
	pass
```

Before: (on a production build)
```
43689 took 2ms
43690 took 4467ms
43691 took 1ms
```
After: (on a debug build)
```
43689 took 9ms
43690 took 9ms
43691 took 9ms
```
(And no this didn't just move where this happens)

---

Allowed array sizes:
Previously (unpacked) arrays with 100 million + elements would fail to allocate / crash
Now all Arrays can have 2 billion elements (if you have the 50GB ram needed for that) and all failures will correctly return error codes.
Unfortunately allocating arrays with over 4 billion elements will appear to work, but the size actually gets truncated to 32bit and thus the result is actually smaller. This happens at the gdscript to native boundary tho, so I am not sure if this can easily be fixed without breaking lots of things or actually making the index/size 64 bit large.

---

Potential issues:
- This touches a pretty core element of the engine
- This tries to not overallocate memory if possible, so buffer overruns are more likely to overrun the allocation entirely
- It is no longer possible to marshal all PackedArray instances into C# arrays, but this issue previously already existed into the other direction.
- ....

Potential advantages:
- Allows bigger arrays
- Paves the way for even bigger arrays
- Probably decreases the overall memory use a bit
- The whole code is no longer super weird
- ....