@Ivorforce 2024-12-22
Awesome! I just found this randomly. Practically, I think it's unavoidable that we have an explicit _capacity_ for `CowData` sooner or later. Boundary reallocations aside, even just having `reserve` on `Vector` would help in some spots.

@RedworkDE Are you still interested in maintaining this PR? I know it's been quite a while since it's been opened.